php73: Backport mysqlnd build fix from PHP 7.4
With newer binutils, all linked in libraries must be explicitly
given. The zlib linkage was not added to mysqlnd resulting in
this build error:

recipe-sysroot-native/usr/lib/libz.so.1: error adding symbols:
DSO missing from command line

Backport the patch that detects zlib and adds it to the mysqlnd
module build.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>